+++ iroh-base/Cargo.toml
@@ -28,6 +28,7 @@ thiserror = "1"
 
 # key module
 aead = { version = "0.5.2", features = ["bytes"], optional = true }
+derive_more = { version = "1.0.0-beta.1", features = ["debug", "display"], optional = true }
 ed25519-dalek = { version = "2.0.0", features = ["serde", "rand_core"], optional = true }
 once_cell = { version = "1.18.0", optional = true }
 rand = { version = "0.8", optional = true }
@@ -36,6 +37,7 @@ ssh-key = { version = "0.6.0", features = ["ed25519", "std", "rand_core"], optio
 ttl_cache = { version = "0.5.1", optional = true }
 crypto_box = { version = "0.9.1", features = ["serde", "chacha20"], optional = true }
 zeroize = { version = "1.5", optional = true }
+url = { version = "2.5.0", features = ["serde"], optional = true }
 
 [dev-dependencies]
 iroh-test = { path = "../iroh-test" }
@@ -48,5 +50,5 @@ default = ["hash", "base32"]
 hash = ["bao-tree", "multibase", "data-encoding", "postcard"]
 base32 = ["data-encoding"]
 redb = ["dep:redb"]
-key = ["dep:ed25519-dalek", "dep:once_cell", "dep:rand", "dep:rand_core", "dep:ssh-key", "dep:ttl_cache", "dep:aead", "dep:crypto_box", "dep:zeroize"]
+key = ["dep:ed25519-dalek", "dep:once_cell", "dep:rand", "dep:rand_core", "dep:ssh-key", "dep:ttl_cache", "dep:aead", "dep:crypto_box", "dep:zeroize", "dep:url", "dep:derive_more"]
