+++ iroh-base/src/ticket/blob.rs
@@ -1,30 +1,41 @@
-//! Tickets for nodes.
-
+//! Tickets for blobs.
 use std::str::FromStr;
 
 use anyhow::{ensure, Result};
-use iroh_base::ticket::{self, Ticket};
+use crate::{
+    hash::{BlobFormat, Hash},
+    ticket::{self, Ticket},
+};
 use serde::{Deserialize, Serialize};
 
-use crate::NodeAddr;
+use crate::node_addr::NodeAddr;
 
 /// A token containing everything to get a file from the provider.
 ///
 /// It is a single item which can be easily serialized and deserialized.
 #[derive(Debug, Clone, PartialEq, Eq, derive_more::Display)]
 #[display("{}", Ticket::serialize(self))]
-pub struct NodeTicket {
+pub struct BlobTicket {
+    /// The provider to get a file from.
     node: NodeAddr,
+    /// The format of the blob.
+    format: BlobFormat,
+    /// The hash to retrieve.
+    hash: Hash,
 }
 
-/// Wire format for [`NodeTicket`].
+/// Wire format for [`BlobTicket`].
+///
+/// In the future we might have multiple variants (not versions, since they
+/// might be both equally valid), so this is a single variant enum to force
+/// postcard to add a discriminator.
 #[derive(Serialize, Deserialize)]
 enum TicketWireFormat {
-    Variant0(NodeTicket),
+    Variant0(BlobTicket),
 }
 
-impl Ticket for NodeTicket {
-    const KIND: &'static str = "node";
+impl Ticket for BlobTicket {
+    const KIND: &'static str = "blob";
 
     fn to_bytes(&self) -> Vec<u8> {
         let data = TicketWireFormat::Variant0(self.clone());
@@ -41,46 +52,67 @@ impl Ticket for NodeTicket {
     }
 }
 
-impl FromStr for NodeTicket {
+impl FromStr for BlobTicket {
     type Err = ticket::Error;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        ticket::Ticket::deserialize(s)
+        Ticket::deserialize(s)
     }
 }
 
-impl NodeTicket {
+impl BlobTicket {
     /// Creates a new ticket.
-    pub fn new(node: NodeAddr) -> Result<Self> {
+    pub fn new(node: NodeAddr, hash: Hash, format: BlobFormat) -> Result<Self> {
         ensure!(!node.info.is_empty(), "addressing info cannot be empty");
-        Ok(Self { node })
+        Ok(Self { hash, format, node })
+    }
+
+    /// The hash of the item this ticket can retrieve.
+    pub fn hash(&self) -> Hash {
+        self.hash
     }
 
     /// The [`NodeAddr`] of the provider for this ticket.
     pub fn node_addr(&self) -> &NodeAddr {
         &self.node
     }
+
+    /// The [`BlobFormat`] for this ticket.
+    pub fn format(&self) -> BlobFormat {
+        self.format
+    }
+
+    /// True if the ticket is for a collection and should retrieve all blobs in it.
+    pub fn recursive(&self) -> bool {
+        self.format.is_hash_seq()
+    }
+
+    /// Get the contents of the ticket, consuming it.
+    pub fn into_parts(self) -> (NodeAddr, Hash, BlobFormat) {
+        let BlobTicket { node, hash, format } = self;
+        (node, hash, format)
+    }
 }
 
-impl Serialize for NodeTicket {
+impl Serialize for BlobTicket {
     fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
         if serializer.is_human_readable() {
             serializer.serialize_str(&self.to_string())
         } else {
-            let NodeTicket { node } = self;
-            (node).serialize(serializer)
+            let BlobTicket { node, format, hash } = self;
+            (node, format, hash).serialize(serializer)
         }
     }
 }
 
-impl<'de> Deserialize<'de> for NodeTicket {
+impl<'de> Deserialize<'de> for BlobTicket {
     fn deserialize<D: serde::Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
         if deserializer.is_human_readable() {
             let s = String::deserialize(deserializer)?;
             Self::from_str(&s).map_err(serde::de::Error::custom)
         } else {
-            let peer = Deserialize::deserialize(deserializer)?;
-            Self::new(peer).map_err(serde::de::Error::custom)
+            let (peer, format, hash) = Deserialize::deserialize(deserializer)?;
+            Self::new(peer, hash, format).map_err(serde::de::Error::custom)
         }
     }
 }
@@ -91,16 +123,19 @@ mod tests {
     use iroh_test::{assert_eq_hex, hexdump::parse_hexdump};
 
     use crate::key::{PublicKey, SecretKey};
-    use std::net::{Ipv4Addr, SocketAddr};
+    use std::net::SocketAddr;
 
     use super::*;
 
-    fn make_ticket() -> NodeTicket {
+    fn make_ticket() -> BlobTicket {
+        let hash = Hash::new(b"hi there");
         let peer = SecretKey::generate().public();
-        let addr = SocketAddr::from((Ipv4Addr::LOCALHOST, 1234));
+        let addr = SocketAddr::from_str("127.0.0.1:1234").unwrap();
         let derp_url = None;
-        NodeTicket {
+        BlobTicket {
+            hash,
             node: NodeAddr::from_parts(peer, derp_url, vec![addr]),
+            format: BlobFormat::HashSeq,
         }
     }
 
@@ -108,7 +143,7 @@ mod tests {
     fn test_ticket_postcard() {
         let ticket = make_ticket();
         let bytes = postcard::to_stdvec(&ticket).unwrap();
-        let ticket2: NodeTicket = postcard::from_bytes(&bytes).unwrap();
+        let ticket2: BlobTicket = postcard::from_bytes(&bytes).unwrap();
         assert_eq!(ticket2, ticket);
     }
 
@@ -116,32 +151,32 @@ mod tests {
     fn test_ticket_json() {
         let ticket = make_ticket();
         let json = serde_json::to_string(&ticket).unwrap();
-        let ticket2: NodeTicket = serde_json::from_str(&json).unwrap();
+        let ticket2: BlobTicket = serde_json::from_str(&json).unwrap();
         assert_eq!(ticket2, ticket);
     }
 
     #[test]
     fn test_ticket_base32() {
+        let hash =
+            Hash::from_str("0b84d358e4c8be6c38626b2182ff575818ba6bd3f4b90464994be14cb354a072")
+                .unwrap();
         let node_id =
             PublicKey::from_str("ae58ff8833241ac82d6ff7611046ed67b5072d142c588d0063e942d9a75502b6")
                 .unwrap();
 
-        let ticket = NodeTicket {
-            node: NodeAddr::from_parts(
-                node_id,
-                Some("http://derp.me./".parse().unwrap()),
-                vec!["127.0.0.1:1024".parse().unwrap()],
-            ),
+        let ticket = BlobTicket {
+            node: NodeAddr::from_parts(node_id, None, vec![]),
+            format: BlobFormat::Raw,
+            hash,
         };
-        let base32 = base32::parse_vec(ticket.to_string().strip_prefix("node").unwrap()).unwrap();
+        let base32 = base32::parse_vec(ticket.to_string().strip_prefix("blob").unwrap()).unwrap();
         let expected = parse_hexdump("
-            00 # variant
+            00 # discriminator for variant 0
             ae58ff8833241ac82d6ff7611046ed67b5072d142c588d0063e942d9a75502b6 # node id, 32 bytes, see above
-            01 # derp url present
-            10 687474703a2f2f646572702e6d652e2f # derp url, 16 bytes, see above
-            01 # one direct address
-            00 # ipv4
-            7f000001 8008 # address, see above
+            00 # derp url
+            00 # number of addresses (0)
+            00 # format (raw)
+            0b84d358e4c8be6c38626b2182ff575818ba6bd3f4b90464994be14cb354a072 # hash, 32 bytes, see above
         ").unwrap();
         assert_eq_hex!(base32, expected);
     }
