+++ iroh-net/src/ticket.rs
@@ -1,6 +1,2 @@
 //! Tickets supported by iroh-net
-mod blob;
-mod node;
-
-pub use blob::BlobTicket;
-pub use node::NodeTicket;
+pub use iroh_base::ticket::NodeTicket;
