+++ iroh/Cargo.toml
@@ -28,7 +28,7 @@ genawaiter = { version = "0.99", default-features = false, features = ["futures0
 hashlink = "0.8.4"
 hex = { version = "0.4.3" }
 iroh-bytes = { version = "0.12.0", path = "../iroh-bytes", features = ["downloader"] }
-iroh-base = { version = "0.12.0", path = "../iroh-base" }
+iroh-base = { version = "0.12.0", path = "../iroh-base", features = ["key"] }
 iroh-io = { version = "0.4.0", features = ["stats"] }
 iroh-metrics = { version = "0.12.0", path = "../iroh-metrics", optional = true }
 iroh-net = { version = "0.12.0", path = "../iroh-net" }
